chore(core/modules): move ws manager to new sub directory
