+++ apps/core/src/modules/process/manager.ts
@@ -5,7 +5,7 @@ import type {
 } from '@fxmanager/shared/types';
 import { LogBuffer } from '../buffer/manager';
 import { ConfigManager } from '../config/manager';
-import { wsManager } from '../ws.manager';
+import { wsManager } from '../ws/manager';
 import { resourceManager } from '../resource/manager';
 
 export class ProcessManager {
